feat: 카드 입력 로직을 useCardInput 훅으로 분리하여 코드 구조 개선
